Change behaviour of OS.Dir.create.
Do not try to chmod existing directories to honour the [mode] and
return a boolean indicating whether the directory already existed (in
which case it can't be assumed to be in [mode]).

This is because we want to be able to easily use OS.Dir.create to
ensure that directory exists before writing to it but the process
may not be able to chmod it (e.g. /tmp).