Fix pipelines (OS.Cmd.out_run_in) on Windows (#51).
When piping two processes we closed on exec the reading pipe before
lauching the writing process. We didn't however clear the close on
exec before lauching the reading process so it would get a closed
reading pipe end.

This can't be seen on POSIX because Unix.create_process is implemented
as fork+exec and standard descriptors are dup'ed for peforming the fd
redirections.